+++ tests/test_optimizer.py
@@ -66,10 +66,10 @@ from desc.optimize import (
     ProximalProjection,
     fmin_auglag,
     fmintr,
+    generic_sgd,
     lsq_auglag,
     lsqtr,
     optimizers,
-    sgd,
 )
 from desc.optimize.optimizer import _parse_x_scale
 from desc.utils import get_all_instances
@@ -252,13 +252,50 @@ class TestSGD:
 
     @pytest.mark.unit
     def test_sgd_convex(self):
-        """Test minimizing convex test function using stochastic gradient descent."""
+        """Test minimizing convex test function using sgd with momentum."""
         x0 = np.ones(2)
 
-        out = sgd(
+        out = generic_sgd(
             scalar_fun,
             x0,
             scalar_grad,
+            method="sgd",
+            verbose=3,
+            ftol=0,
+            xtol=0,
+            gtol=1e-12,
+            maxiter=2000,
+        )
+        np.testing.assert_allclose(out["x"], SCALAR_FUN_SOLN, atol=1e-4, rtol=1e-4)
+
+    @pytest.mark.unit
+    def test_adam_convex(self):
+        """Test minimizing convex test function using ADAM."""
+        x0 = np.ones(2)
+
+        out = generic_sgd(
+            scalar_fun,
+            x0,
+            scalar_grad,
+            method="adam",
+            verbose=3,
+            ftol=0,
+            xtol=0,
+            gtol=1e-12,
+            maxiter=2000,
+        )
+        np.testing.assert_allclose(out["x"], SCALAR_FUN_SOLN, atol=1e-4, rtol=1e-4)
+
+    @pytest.mark.unit
+    def test_rmsprop_convex(self):
+        """Test minimizing convex test function using RMSProp."""
+        x0 = np.ones(2)
+
+        out = generic_sgd(
+            scalar_fun,
+            x0,
+            scalar_grad,
+            method="rmsprop",
             verbose=3,
             ftol=0,
             xtol=0,